+++ src/smallestai/waves/ops/raw_client.py
@@ -38,7 +38,7 @@ class RawOpsClient:
         """
         _response = self._client_wrapper.httpx_client.request(
             "waves/v1/health",
-            base_url=self._client_wrapper.get_environment().atoms,
+            base_url=self._client_wrapper.get_environment().waves,
             method="GET",
             request_options=request_options,
         )
@@ -97,7 +97,7 @@ class AsyncRawOpsClient:
         """
         _response = await self._client_wrapper.httpx_client.request(
             "waves/v1/health",
-            base_url=self._client_wrapper.get_environment().atoms,
+            base_url=self._client_wrapper.get_environment().waves,
             method="GET",
             request_options=request_options,
         )